### 2018-09-29 #986 @whedon
[PRE REVIEW]: fibergen: An introductory tool for FFT-based material homogenization
**Submitting author:** @fospald (<a href="http://orcid.org/0000-0001-8372-9179">Felix Ospald</a>)
**Repository:** <a href="https://github.com/fospald/fibergen" target ="_blank">https://github.com/fospald/fibergen</a>
**Version:** 2018.1
