+++ joss/paper.md
@@ -1,0 +1,86 @@
+---
+title: '``fibergen``: An introductory tool for FFT-based material homogenization'
+tags:
+  - material homogenization
+  - FFT
+  - Python
+  - C++
+  - heat equation
+  - Stokes flow
+  - Darcy flow
+  - linear elasticity
+  - hyperelasticity
+authors:
+  - name: Felix Ospald
+    orcid: 0000-0001-8372-9179
+    affiliation: "1"
+affiliations:
+ - name: Research Group Numerical Mathematics, TU Chemnitz, Chemnitz, Germany
+   index: 1
+date: 2 October 2018
+bibliography: paper.bib
+---
+
+# Summary
+
+Many engineering applications require material properties of heterogeneous media. For instance elasticity constants and thermal conductivity of fiber reinforced plastics, the viscosity of a suspension, permeability of porous media, electrical conductivity of sintered metals, ...
+Laboratory experiments for determining these properties are often difficult, expensive and error-prone.
+Therefore, increasingly computer simulations are used instead, both for the computation of effective properties and the generation of the geometry of such materials in order to avoid the use of expensive computer tomography (CT) scans.
+
+[``fibergen``](http://fospald.github.io/fibergen/) is a easy to use tool (command-line/GUI) for the homogenization (i.e. obtaining effective material parameters of an equivalent homogeneous media) and generation of periodic materials, especially but not limited to fiber reinforced composites produced by injection molding.
+The voxel-based spectral method relies on the fast Fourier transform (FFT) and OpenMP parallelization for highly efficient computations, which can be carried out on an average desktop computer. The tool also includes a geometry generator, but is also capable to process raw data (e.g. CT data).
+
+# Relation to similar projects
+
+There exists a number of similar projects for the homogenization of materials, developed by well-known people in this research area:
+* [AMITEX](http://www.maisondelasimulation.fr/projects/amitex/html), Lionel Gélébart, 
+Atomic Energy and Alternative Energies Commission, France, (Fortran)
+* [morphhom](http://cmm.ensmp.fr/morphhom), François Willot, Université Paris, France, (Fortran/C)
+* [DAMASK](http://damask.mpie.de/Documentation/SpectralSolver), Martin Diehl, Max-Planck-Institut für Eisenforschung, Germany, (Fortran, C, C++)
+* [janus](http://github.com/sbrisard/janus), Sébastien Brisard, Université Paris, France, (Cython)
+* [FFTHomPy](http://github.com/vondrejc/FFTHomPy), Jaroslav Vondřejc, TU Braunschweig, Germany (Python)
+
+All these projects are written in different programming languages and implement different solver methods and provide different features. Most of these tools provide only the pure solver without a geometry generator and without post-processing (visualization) of the results.
+``fibergen`` on the other hand provides a solver but also an GUI/IDE for editing project files, running the solver and post-processing.
+It is also written in object-oriented C++ and provides a simple Python interface for the automation of computations.
+
+# Features
+
+``fibergen`` implements FFT-based homogenization based on the Lippmann-Schwinger equation.
+It includes the original method by [@MoulinecSuquet1994] and the more recent discretizations [@Willot2015] and [@SchneiderOspaldKabel2015]. Especially the staggered grid method is not implemented by any other of the listed tools, although it is the superior method. ``fibergen`` also implements efficient fixed point and Newton-Krylov solvers for FFT-based homogenization of elasticity at large deformations as described in [@KabelBöhlkeSchneider2014].
[…]
