+++ joss/paper.bib
@@ -1,0 +1,152 @@
+@misc{Jones2001,
+	author = {Jones, Eric and Oliphant, Travis and Peterson, Pearu and Others},
+	title = {SciPy: Open source scientific tools for Python},
+	url = {http://www.scipy.org/},
+	year = {2001}
+}
+@book{Oliphant2006,
+	title={A guide to NumPy},
+	author={Oliphant, Travis E},
+	volume={1},
+	year={2006},
+	publisher={Trelgol Publishing USA}
+}
+@ARTICLE{Montgomery-SmithJackSmith2011,
+  title = {The Fast Exact Closure for Jeffery’s equation with diffusion},
+  journal = {Journal of Non-Newtonian Fluid Mechanics},
+  volume = {166},
+  number = {7-8},
+  pages = {343-353},
+  year = {2011},
+  note = {},
+  issn = {0377-0257},
+  doi = {http://dx.doi.org/10.1016/j.jnnfm.2010.12.010},
+  url = {http://www.sciencedirect.com/science/article/pii/S0377025711000048},
+  author = {Stephen Montgomery-Smith and David Jack and Douglas E. Smith},
+  keywords = {Directional orientation},
+  keywords = {Rheological properties},
+  keywords = {Injection molding},
+  keywords = {Jeffery’s equation with rotary diffusion},
+	file	= {555:Montgomery-Smith_Jack_Smith__The_Fast_Exact_Closure_for_Jeffery’s_equation_with_diffusion.pdf:PDF}
+}
+@ARTICLE{KabelBöhlkeSchneider2014,
+  year={2014},
+  issn={0178-7675},
+  journal={Computational Mechanics},
+  volume={54},
+  number={6},
+  doi={10.1007/s00466-014-1071-8},
+  title={Efficient fixed point and Newton-Krylov solvers for FFT-based homogenization of elasticity at large deformations},
+  publisher={Springer Berlin Heidelberg},
+  keywords={Composite materials; Finite deformations; Lippmann-Schwinger equation; FFT; 74B20; 45G10; 65T50},
+  author={Kabel, Matthias and Böhlke, Thomas and Schneider, Matti},
+  pages={1497-1514},
+  language={English},
+	file	= {555:Kabel_Böhlke_Schneider__Efficient_fixed_point_and_Newton-Krylov_solvers_for_FFT-based_homogenization_of_elasticity_at_large_deformations.pdf:PDF}
+}
+@ARTICLE{KabelMerkertSchneider2014,
+  title = {{FFT-based homogenization using composite voxels}},
+  journal = {Computer Methods in Applied Mechanics and Engineering},
+  volume = {},
+  number = {},
+  pages = {},
+  year = {2014},
+  note = {},
+  author = {M. Kabel and D. Merkert and M. Schneider},
+  doi={},
+	file	= {555:Kabel_Merkert_Schneider__FFT-based_homogenization_using_composite_voxels.pdf:PDF}
+}
+@ARTICLE{Kabel2016,
+  author="Kabel, Matthias and Fliegener, Sascha and Schneider, Matti",
+  title="Mixed boundary conditions for FFT-based homogenization at finite strains",
+  journal="Computational Mechanics",
+  year="2016",
+  month="Feb",
+  day="01",
+  volume="57",
+  number="2",
+  pages="193--210",
+  issn="1432-0924",
+  doi="10.1007/s00466-015-1227-1",
+	file	= {555:Kabel__Mixed_boundary_conditions_for_FFT-based_homogenization_at_finite_strains.pdf:PDF}
+}
+@article {SchneiderOspaldKabel2015,
+	author = {Schneider, Matti and Ospald, Felix and Kabel, Matthias},
[…]
