### 2018-10-05 500311d
updated paper
