+++ joss/paper.bib
@@ -20,7 +20,7 @@
   year = {2011},
   note = {},
   issn = {0377-0257},
-  doi = {http://dx.doi.org/10.1016/j.jnnfm.2010.12.010},
+  doi = {10.1016/j.jnnfm.2010.12.010},
   url = {http://www.sciencedirect.com/science/article/pii/S0377025711000048},
   author = {Stephen Montgomery-Smith and David Jack and Douglas E. Smith},
   keywords = {Directional orientation},
@@ -55,6 +55,18 @@
   author = {M. Kabel and D. Merkert and M. Schneider},
   doi={},
 	file	= {555:Kabel_Merkert_Schneider__FFT-based_homogenization_using_composite_voxels.pdf:PDF}
+}
+@article{KabelMerkertSchneider2015,
+	title = "Use of composite voxels in FFT-based homogenization",
+	journal = "Computer Methods in Applied Mechanics and Engineering",
+	volume = "294",
+	pages = "168 - 188",
+	year = "2015",
+	issn = "0045-7825",
+	doi = "10.1016/j.cma.2015.06.003",
+	url = "http://www.sciencedirect.com/science/article/pii/S0045782515001954",
+	author = "Matthias Kabel and Dennis Merkert and Matti Schneider",
+	keywords = "Composite materials, Homogenization, Lippmann–Schwinger equation, FFT"
 }
 @ARTICLE{Kabel2016,
   author="Kabel, Matthias and Fliegener, Sascha and Schneider, Matti",
@@ -133,7 +145,7 @@
   year = "2015",
   note = "",
   issn = "1631-0721",
-  doi = "http://dx.doi.org/10.1016/j.crme.2014.12.005",
+  doi = "10.1016/j.crme.2014.12.005",
   url = "http://www.sciencedirect.com/science/article/pii/S1631072114002149",
   author = "FranÃ§ois Willot",
   keywords = "\{FFT\} methods",
