### 2018-10-08 #986 @fospald
@arfon now I have the paper and license added
