### 2018-10-09 #986 @jedbrown
Thanks for adding the paper @fospald and working with Whedon to get it built. I'm currently reaching out to potential reviewers. Additional suggestions would be welcome. Once we have reviewers, the conversation will move to a new REVIEW issue where they will be able to provide feedback and recommendations.
