### 2018-10-18 #986 @jedbrown
@nnadeau @ctdegroot Would you be available to review this submission?
