### 2018-10-18 #986 @nnadeau
@jedbrown unfortunately, I'll have to pass on this review.
