### 2018-10-18 #1027 @whedon
[REVIEW]: fibergen: An introductory tool for FFT-based material homogenization
**Submitting author:** @fospald (<a href="http://orcid.org/0000-0001-8372-9179">Felix Ospald</a>)
**Repository:** <a href="https://github.com/fospald/fibergen" target ="_blank">https://github.com/fospald/fibergen</a>
**Version:** v1.0
