### 2018-10-18 fospald/fibergen#1 @tisaac
A tagged release
Hi @fospald,

One of my checkboxes in the JOSS review (openjournals/joss-reviews#1027) asks whether the version we're reviewing matches the "github release".  fibergen doesn't currently have any releases.  I'm not sure whether a release version is mandatory or not (EDIT: editor has indicated it is not), but either way I would recommend that you start tagging releases.  Numbered releases help in distribution (if you're thinking of packaging fibergen), and they help users keep track of which version they can use (you may make changes down the line that introduce incompatibilities).

Since you're starting fresh, I'd suggest you consider the [semantic versioning](semver.org) convention.
