### 2018-10-18 #1027 @jedbrown
@tisaac A tag should be made after the review (which may include new commits to the software), as part of archiving (using Zenodo or similar). It isn't necessary at this stage.
