### 2018-10-18 #1027 @tisaac
I'm checking off "Does the full list of paper authors seem appropriate and complete?", but the second commit (fospald/fibergen@c6a91c98e) appears to be an import an existing project, so I'm relying on the honesty of the author.
