### 2018-10-18 fospald/fibergen#2 @tisaac
Unable to install: developer local files
https://github.com/fospald/fibergen/blob/207d07d6ae7d7d2ff4d86729ade48b8b26ff77cf/build.sh#L6

Users should not have to modify hard-coded paths in a build script.

At a minimum, please:
- Make the build script depend on environment variables for the locations of dependencies
- Be complete and specific (version numbers) in describing dependencies in installation instructions

For something more sustainable, please consider a flexible [compiling and configuration tool](https://en.wikipedia.org/wiki/Category:Compiling_tools) to help make this project portable.

(Issue raised as part of JOSS review openjournals/joss-reviews#1027)
