### 2018-10-18 fospald/fibergen#2 @fospald
Sorry, I forgot to remove these lines. But it should also work, if the files are not found. The build script uses cmake. Is there any other error message?
Do you mean I should make build.sh portable, e.g. only use cmake without build.sh?
The version requirements are Python 3 and Qt 5, for the other libraries there are no restrictions (it should work with the latest version, but also older versions)
