### 2018-10-18 #1027 @fospald
I'm the original author of every line of the code and all demos, etc.
I did not use github right from the beginning for the project. I'm adding the project because research and publications are done.
