### 2018-10-19 fospald/fibergen#2 @tisaac
Oh, I didn't see that after `build.sh` everything is cmake based.  My mistake, you can ignore my suggestion about configuration tools.

On the other hand, `build.sh` isn't currently parameterized: there's no way to pass arguments to the cmake invocation.  CMake is common enough: maybe you should recommend users just use it directly, and take `build.sh` out of the repo?

You're also right that those lines didn't cause my configure errors.  You should probably still remove them and other references to `/LOCAL/` (Which also appears in `CMakeLists.txt`).
