### 2018-10-22 bb97794
cleaned build script
