+++ CMakeLists.txt
@@ -79,25 +79,16 @@ FIND_PACKAGE(LAPACK REQUIRED)
 INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${LAPACK_INCLUDE_DIR})
 
 # Find FFTW library
-IF(TRUE)
-	SET(USE_FFTWF ON)
-	SET(USE_FFTWD ON)
-	SET(USE_FFTWL OFF)
-	SET(FFTWF_FIND_REQUIRED OFF)
-	SET(FFTWD_FIND_REQUIRED ON)
-	SET(FFTWL_FIND_REQUIRED OFF)
-	FIND_PACKAGE(FFTW REQUIRED)
-	INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${FFTW_INCLUDES})
-	IF(FFTWF_FOUND)
-	    ADD_DEFINITIONS(-DFFTWF_ENABLED)
-	ENDIF()
-ELSE()
-	# TODO: this is a quick hack
-	SET(FFTWF_LIB fftw3f fftw3f_omp)
-	SET(FFTWD_LIB fftw3 fftw3_omp)
-	SET(FFTWL_LIB "")
-	INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} "/LOCAL/Software/FFTW-3.3.4/include")
-	LINK_DIRECTORIES(${LINK_DIRECTORIES} "/LOCAL/Software/FFTW-3.3.4/lib64")
+SET(USE_FFTWF ON)
+SET(USE_FFTWD ON)
+SET(USE_FFTWL OFF)
+SET(FFTWF_FIND_REQUIRED OFF)
+SET(FFTWD_FIND_REQUIRED ON)
+SET(FFTWL_FIND_REQUIRED OFF)
+FIND_PACKAGE(FFTW REQUIRED)
+INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${FFTW_INCLUDES})
+IF(FFTWF_FOUND)
+    ADD_DEFINITIONS(-DFFTWF_ENABLED)
 ENDIF()
 
 SET(SOURCES src/fibergen.cpp)
