### 2018-10-22 9e34c12
updated build script
