+++ build.sh
@@ -4,7 +4,6 @@ HOST=$(uname -n | tr " -" "__")
 FIBERGEN_DIR=$(cd $(dirname "$0") && pwd)
 
 BUILD_DIR="$FIBERGEN_DIR/build/$HOST"
-test -d "$BUILD_DIR" && rm -rf "$BUILD_DIR"
 mkdir -p "$BUILD_DIR" || exit
 cd "$BUILD_DIR" || exit
 cmake -DCMAKE_VERBOSE_MAKEFILE:BOOL=OFF "$@" ../.. || exit
