+++ build.sh
@@ -9,6 +9,7 @@ cd "$BUILD_DIR" || exit
 cmake -DCMAKE_VERBOSE_MAKEFILE:BOOL=OFF "$@" ../.. || exit
 make -j || exit
 
+echo ""
 echo "Build successful!"
 echo "Please update your environmet variables, i.e.:"
 echo "export PATH=\$PATH:$FIBERGEN_DIR/bin"
