### 2018-10-22 fospald/fibergen#2 @fospald
Hi, I have cleaned up the scripts. The build script is simple enough for editing. Now you can also pass CMake parameters as arguments to build.sh. Do you still have errors?
