### 2018-10-22 fospald/fibergen#1 @fospald
Hi @tisaac,
I will make only small changes in the future and try to keep things compatible. As this is not a library for developers, but a tool I use the new ISO versioning currently 2018.1. But I will create a release once we have finished the reviewing process.
