### 2018-10-22 fospald/fibergen#1 @tisaac
Good point about being a tool rather than a library.
