### 2018-10-22 fospald/fibergen#2 @tisaac
The new bulid script works for me, thanks!
