### 2018-10-23 fospald/fibergen#3 @tisaac
Convert tex citations to links in README.md
Could you please turn the `\cite{}` commands in to doi.org links?

(Issue raised as part of JOSS review /openjournals/joss-reviews#1027
