### 2018-10-23 fospald/fibergen#4 @tisaac
Automated testing
By my count there are three demos (elasticity/hashin, elasticity/laminate, viscosity/nunan_keller) where empirical or analytic values that can be compared against the output of the solver are stored as comments in the xml files.  It would be great if these tests could be turned into a `make test` target in CMake as automated tests.

(Issue opened as part of JOSS review openjournals/joss-reviews#1027
