### 2018-10-25 972969a
updated tests
