### 2018-10-25 fospald/fibergen#4 @fospald
I have updated the test routines (need recompile). You can perform some tests using
fibergen --test
The demos are not suitable for testing, as the results depend on the resolution, solver method, etc. or are too simple (laminate) or values too inaccurate (nunan_keller) instead I check some central operations and operators.
