### 2018-10-28 0fa0b23
added solver
