+++ src/fibergen.cpp
@@ -97,6 +97,7 @@ multigrid improvements:
 #include <boost/numeric/bindings/lapack/geev.hpp>
 #include <boost/numeric/bindings/lapack/gesv.hpp>
 #include <boost/numeric/bindings/lapack/syev.hpp>
+#include <boost/numeric/bindings/lapack/sysv.hpp>
 
 #include <boost/accumulators/accumulators.hpp>
 #include <boost/accumulators/statistics/stats.hpp>
@@ -9485,6 +9486,7 @@ public:
 
 		#pragma omp parallel for schedule (static) collapse(2)
 		for (std::size_t j = 0; j < m; j++) {
+			//if (c[j] == 0) continue;
 			for (std::size_t i = 0; i < n; i++) {
 				(*this)[j][i] += c[j];
 			}
@@ -9519,6 +9521,7 @@ public:
 		
 		#pragma omp parallel for schedule (static) collapse(2)
 		for (std::size_t j = 0; j < m; j++) {
+			//if (c[j] == 1) continue;
 			for (std::size_t i = 0; i < n; i++) {
 				(*this)[j][i] *= c[j];
 			}
@@ -9607,6 +9610,8 @@ public:
 	noinline void scale(T s)
 	{
 		Timer __t("scale", false);
+
+		if (s == 1) return;
 
 #ifdef USE_MANY_FFT
 		#pragma omp parallel for schedule (static)
@@ -11485,7 +11490,7 @@ public:
 
 	std::vector<pPhase> phases;
 
-	virtual void getRefMaterial(TensorField<T>& F, T& mu_0, T& lambda_0, bool zero_trace) const = 0;
+	virtual void getRefMaterial(TensorField<T>& F, T& mu_0, T& lambda_0, bool zero_trace, bool polarization) const = 0;
 	virtual T meanW(const TensorField<T>& F) const = 0;
 	virtual T minEig(const TensorField<T>& F, bool zero_trace) const = 0;
 	virtual void eig(std::size_t i, const T* F, T& lambda_min_p, T& lambda_max_p, bool zero_trace) const = 0;
@@ -11544,7 +11549,7 @@ class MixedMaterialLaw : public MixedMaterialLawBase<T, P>
 {
 public:
 	// comupute (lambda_min + lambda_max)/2 for dP/dF
-	noinline void getRefMaterial(TensorField<T>& F, T& mu_0, T& lambda_0, bool zero_trace) const
+	noinline void getRefMaterial(TensorField<T>& F, T& mu_0, T& lambda_0, bool zero_trace, bool polarization) const
 	{
 		Timer __t("getRefMaterial", false);
 
@@ -11616,8 +11621,15 @@ br:
 			lambda_min = 0;
 		}
 
-		mu_0 = 0.5*(lambda_min + lambda_max);
-		lambda_0 = 0.5*(lambda_max - lambda_min);
+		if (polarization) {
+			mu_0 = std::sqrt(lambda_min*lambda_max);
+		}
+		else {
+			mu_0 = 0.5*(lambda_min + lambda_max);
+			//lambda_0 = 0.5*(lambda_max - lambda_min);
+		}
+
+		lambda_0 = 0.0;
 	}
 
 	// mean quantities for W 
@@ -14038,7 +14050,7 @@ protected:
 	std::string _update_ref;	// reference media update interval, "always", "loadstep"
 	std::string _error_estimator;	// method for convergence check (at the moment the basic scheme uses "quick" and cg uses the residual)
[…]
