### 2018-11-02 f6300d0
updated polarization scheme
