+++ src/fibergen.cpp
@@ -10044,6 +10044,42 @@ public:
 		}
 	}
 
+	// Eyre, D. J., & Milton, G. W. (1999). A fast numerical scheme for computing the response of composites using grid refinement. The European Physical Journal Applied Physics, 6(1), 41–47. doi:10.1051/epjap:1999150
+	// compute sigma = (C0 + C)(C0 - C)^{-1} epsilon, C0 = 2*mu_0
+	// if inv is true then: sigma = -(C0 - C)^{-1} epsilon
+	virtual void calcPolarization(std::size_t i, T mu_0, const ublas::vector<T>& F, ublas::vector<T>& P,
+		std::size_t dim, bool inv) const
+	{
+		ublas::matrix<T> C(dim, dim), C2;
+
+		// note: for linear problems independent of F
+		const T* Id;
+		if (dim == 3) {
+			Id = TensorIdentity<T,3>::Id;
+		}
+		else if (dim == 6) {
+			Id = TensorIdentity<T,6>::Id;
+		}
+		else {
+			Id = TensorIdentity<T,9>::Id;
+		}
+
+		this->dPK1(i, F.data().begin(), 1, false, Id, C.data().begin(), dim);
+
+		// L0 = 2*mu_0
+		// C1 = C - 2*mu_0*ublas::identity_matrix<T>(dim);
+		C2 = C + 2.0*mu_0*ublas::identity_matrix<T>(dim);
+
+		// Solve C2*Q = F
+		P = F;
+		lapack::gesv(C2, P);
+
+		if (!inv) {
+			// P = C1*C2inv*F
+			P = ublas::prod(C, P) - 2.0*mu_0*P;
+		}
+	}
+
 	virtual std::string str() const = 0;
 
 	virtual void readSettings(const ptree::ptree& pt) { }
@@ -10893,6 +10929,49 @@ public:
 		}
 	}
 
+	void calcPolarization(std::size_t i, T mu_0, const ublas::vector<T>& F, ublas::vector<T>& P,
+		std::size_t dim, bool inv) const
+	{
+		// C = 2*mu*Id + lambda*I*I
+
+		// L0 = 2*mu_0
+		// C1 = 2*(mu-mu_0)*Id + lambda*II
+		// C2 = 2*(mu+mu_0)*Id + lambda*II
+
+		// The inverse of C2 is:
+		// Simplify[Inverse[{{2*mu + L, L, L, 0, 0, 0}, {L, 2*mu + L, L, 0, 0, 0}, {L, L, 2*mu + L, 0, 0, 0}, {0, 0, 0, 2*mu, 0, 0}, {0, 0, 0, 0, 2*mu, 0}, {0, 0, 0, 0, 0, 2*mu}}]]
+		// Simplify[(L + mu)/(3 L mu + 2 mu^2) - 1/(2 mu)]
+		// inv(C2) = 1/(2*m)*Id - (lambda/(2*m*(3*lambda + 2*m)))*II
+		// where m = mu+mu_0
+
+		// Solve C2*P = F
+
+		T m = 2.0*(mu+mu_0);
+		T a = 1.0/m;
+		T b = lambda/(m*(3.0*lambda + m));
+		T tr_F = F[0] + F[1] + F[2];
+
+		P[0] = a*F[0] - b*tr_F;
+		P[1] = a*F[1] - b*tr_F;
+		P[2] = a*F[2] - b*tr_F;
+		P[3] = a*F[3];
+		P[4] = a*F[4];
+		P[5] = a*F[5];
+
+		if (!inv) {
+			// P = C1*C2inv*F
+			m = 2.0*(mu-mu_0);
[…]
