### 2018-11-09 #1027 @ctdegroot
@jedbrown Working on the review. Hope to finish next week.

@fospald I am supposed to check "Does the release version given match the GitHub release (2018.1)?", but the GitHub repo shows no releases tagged. Can you clarify this for me?
