### 2018-11-09 #1027 @ctdegroot
@fospald I read the review above and can see that the release issue has already been discussed.
