### 2018-11-09 #1027 @ctdegroot
_Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._ I don't see any examples of how to use the code. If there are some, can you please tell me where they are and please put something in the README about these.

_Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_ I don't see any guidelines for how third parties can contribute to the software. Can you add something to the README?
