### 2018-11-12 #1027 @fospald
> @fospald I am supposed to check "Does the release version given match the GitHub release (2018.1)?", but the GitHub repo shows no releases tagged. Can you clarify this for me?

I will make only minor changes to the project and keep everything compatible. I don't want to maintain several branches. There will be no future releases with big changes.
