### 2018-11-12 #1027 @fospald
> _Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)._ I don't see any examples of how to use the code. If there are some, can you please tell me where they are and please put something in the README about these.

You run the GUI and then you can select one of the demos (located in the demo directory).
[See here](https://fospald.github.io/fibergen/page/images/screenshot_1.png)
The editor gives help about the available settings in the project file.
[See here](https://fospald.github.io/fibergen/page/images/screenshot_2.png)

> _Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_ I don't see any guidelines for how third parties can contribute to the software. Can you add something to the README?

Ok. I added a "Contributing" section in the README.
