### 2018-11-12 #1027 @fospald
> I cannot find any documentation of the code API. It seems that all of the core code is stuffed into the 27,000 line fibergen.cpp. How would one go about understanding the code and potentially working with it, without the GUI? Maybe try Doxygen to at least give a basic bit of documentation?

@ctdegroot End users should use the GUI and should not do any coding.
But I understand the issue. I will try my best add some documentation with Doxygen and reply when I'm done. Also I know that the single file concept is not ideal. Limited time for the project...
