+++ src/fibergen.cpp
@@ -1,3 +1,10 @@
+/**
+
+\brief fibergen
+
+
+
+*/
 
 // http://de.mathworks.com/company/newsletters/articles/the-watershed-transform-strategies-for-image-segmentation.html
 
@@ -241,7 +248,7 @@ std::ostream& operator<<(std::ostream& os, const TTYOnly& tto)
 #endif
 
 
-// Class for logging to cout
+//! Class for logging
 class Logger
 {
 protected:
@@ -289,13 +296,13 @@ public:
 		_stream.reset(new std::ostream(_tee_stream->rdbuf()));
 	}
 
-	// increase indent
+	//! Increase indent
 	void incIndent() { _indent++; }
 
-	// decrease indent
+	//! Decrease indent
 	void decIndent() { _indent = (size_t) std::max(((int)_indent)-1, 0); }
 
-	// write indent to stream
+	//! Write indent to stream
 	void indent(std::ostream& stream) const
 	{
 		std::string space(2*_indent, ' ');
@@ -312,7 +319,7 @@ public:
 		_stream->flush();
 	}
 
-	// return cout stream
+	//! Return standard output stream
 	std::ostream& cout()
 	{
 		std::ostream& stream = *_stream;
@@ -322,7 +329,7 @@ public:
 		return stream;
 	}
 
-	// return cerr stream
+	//! Return error stream
 	std::ostream& cerr() const
 	{
 		std::ostream& stream = std::cerr;
@@ -331,7 +338,7 @@ public:
 		return stream;
 	}
 
-	// return static instance
+	//! Return static instance
 	static Logger& instance()
 	{
 		if (!_instance) {
@@ -342,14 +349,18 @@ public:
 	}
 };
 
+// Static logger instance
 boost::shared_ptr<Logger> Logger::_instance;
 
+// Shortcut macros for cout and cerr
 #define LOG_COUT Logger::instance().cout()
 #define LOG_CERR Logger::instance().cerr()
 
-
+// Static exception object
 boost::shared_ptr<std::exception> _except;
 
+
+//! Set current exception message and print message to cerr
 void set_exception(const std::string& msg)
 {
 	#pragma omp critical
@@ -362,8 +373,7 @@ void set_exception(const std::string& msg)
 	}
 }
 
-
-
+//! Print backtrace of current function calls
 inline void print_stacktrace(std::ostream& stream)
 {
 	// print stack trace
@@ -433,10 +443,11 @@ inline void print_stacktrace(std::ostream& stream)
 
 
 
-
+// Static empty ptree object
 static ptree::ptree empty_ptree;
 
 
+//! Open file for output, truncate file if already exists
 inline void open_file(std::ofstream& fs, const std::string& filename)
 {
 	fs.open(filename.c_str(), std::ofstream::out | std::ofstream::trunc | std::ofstream::binary);
@@ -447,9 +458,13 @@ inline void open_file(std::ofstream& fs, const std::string& filename)
 }
 
 
+//! Class with math operations for Voigt's notation
 class Voigt
 {
 public:
+	//! Identity matrix for dimension dim
+	//! \param dim the dimension (3, 6 or 9)
+	//! \return matrix
 	template <typename T>
 	static inline ublas::matrix<T> Id4(std::size_t dim) 
 	{
@@ -464,6 +479,9 @@ public:
 		return Id;
 	}
 
+	//! Outer product of rank 2 identity matrices as Voigt matrix
+	//! \param dim the dimension (3, 6 or 9)
+	//! \return matrix
 	template <typename T>
 	static inline ublas::matrix<T> II4(std::size_t dim) 
 	{
@@ -474,6 +492,8 @@ public:
[…]
