+++ doc/doxygen/fibergen.cpp
@@ -1,0 +1,1 @@
+../../src/fibergen.cpp
