+++ doc/doxygen/Makefile
@@ -1,0 +1,12 @@
+
+DOXYFILE:=Doxyfile
+
+DOXYGEN:=doxygen
+
+all: doxygen
+
+doxygen: $(DOXYFILE)
+	$(DOXYGEN) $<
+
+clean:
+	rm -rf html latex
