### 2018-11-13 b93fd70
updated documentation
