+++ src/fibergen.cpp
@@ -1716,8 +1716,8 @@ class TetVTKReader
 public:
 	//! Read data from file
 	//! \param filename the VTK filename
-	//! \param points
-	//! \param tets
+	//! \param points output of the vertices
+	//! \param tets output of vector of 4-tupes describing the vertex indices of each tetrahedron
 	void read(const std::string& filename, std::vector< ublas::c_vector<T,3> >& points, std::vector< ublas::c_vector<std::size_t,4> >& tets)
 	{
 		/*
@@ -1848,6 +1848,7 @@ public:
 
 
 
+//! Class for reading a tetrahedron mesh from a Dolfin/FEniCS XLM mesh file
 template <typename T>
 class TetDolfinXMLReader
 {
@@ -1894,6 +1895,7 @@ public:
 };
 
 
+//! Class for reading a triangle surface mesh from a STL file
 template <typename T>
 class STLReader
 {
@@ -1967,7 +1969,7 @@ public:
 	}
 };
 
-// Standard normal (mu=0, sigma=1) distributed number generator
+//! Standard normal (mu=0, sigma=1) distributed number generator
 template<typename T>
 class RandomNormal01
 {
@@ -1987,16 +1989,17 @@ public:
 	{
 	}
 	
+	//! change random seed
 	void seed(int s) {
 		// http://stackoverflow.com/questions/4778797/setting-seed-boostrandom
 		_rnd.engine().seed(s);
 		_rnd.distribution().reset();
 	}
 	
-	// return random number
+	//! return random number
 	T rnd() { return _rnd(); }
 	
-	// return static instance
+	//! return static instance
 	static RandomNormal01& instance() { return _instance; }	
 };
 
@@ -2004,7 +2007,7 @@ template<typename T>
 RandomNormal01<T> RandomNormal01<T>::_instance;
 
 
-// Uniform number generator in [0,1]
+//! Uniform number generator in [0,1]
 template<typename T>
 class RandomUniform01
 {
@@ -2024,16 +2027,17 @@ public:
 	{
 	}
 	
+	//! change random seed
 	void seed(int s) {
 		// http://stackoverflow.com/questions/4778797/setting-seed-boostrandom
 		_rnd.engine().seed(s);
 		_rnd.distribution().reset();
 	}
 	
-	// return random number
+	//! return random number
 	T rnd() { return _rnd(); }
 	
-	// return static instance
+	//! return static instance
 	static RandomUniform01& instance() { return _instance; }	
 };
 
@@ -2082,6 +2086,10 @@ void stable_A(ublas::c_vector<T, DIM>& a)
 }
 
 
+
+//! Moments of angular central Gaussian distribution from distribution parameters
+//! \param b input ACG parameters
+//! \param a output moments
 template<typename T, int DIM>
 void A_from_B(const ublas::c_vector<T, DIM>& b, ublas::c_vector<T, DIM>& a)
 {
@@ -2091,6 +2099,13 @@ void A_from_B(const ublas::c_vector<T, DIM>& b, ublas::c_vector<T, DIM>& a)
 }
 
 
+//! Inversion of moments of angular central Gaussian distribution using fixed point method
+//! \param a input moments
+//! \param b output ACG parameters
+//! \param tol tolerance
+//! \param max_iter maximum number of iterations
+//! \param step step size
+//! \return number of required iterations
 template<typename T, int DIM>
 std::size_t compute_B_from_A_fixedpoint(const ublas::c_vector<T, DIM>& a, ublas::c_vector<T, DIM>& b, T tol, std::size_t max_iter = 1000000, T step = 0.5)
 {
[…]
