### 2018-11-13 #1027 @fospald
@ctdegroot I added a Doxygen documentation. You have to
git pull
cd doc/doxygen
make (or just doxygen)
firefox html/index.html
(also added this to the README.md)
