### 2018-11-15 #1027 @ctdegroot
> > @fospald Where are the test cases located? To what extent do they cover the code?
>
> @ctdegroot The test cases are located in fibergen.cpp at line 23113 to 24194.
> The tests cover core math operations, functions to compute geometric intersections, material laws (derivatives), solver operators for various methods.

@fospald Can you pull the test cases out of this file? It would be a lot easier for users to learn the code from the test cases if they are easier to find.
