### 2018-11-15 #1027 @ctdegroot
> @ctdegroot I added a Doxygen documentation. You have to
> git pull
> cd doc/doxygen
> make (or just doxygen)
> firefox html/index.html
> (also added this to the README.md)

Ok, this helps. I don't think the code structure is ideal by any means. It would be much better to create different header files and source files for different purposes. It's acceptable for the review though. I won't insist on this.
