+++ README.md
@@ -25,11 +25,14 @@ A FFT-based homogenization tool.
 The following libraries are required, which are likely already installed on your system:
 * [CMake](https://cmake.org/)
 * [gcc](https://gcc.gnu.org/)
+* [OpenMP](https://www.openmp.org/)
 * [boost](https://www.boost.org/) incl. boost.python 3 and [boost-numeric-bindings](https://mathema.tician.de/software/boost-numeric-bindings/)
+* [FFTW3](http://www.fftw.org/) library
+* [lapack](www.netlib.org/lapack/) library
+* [libpng](http://www.libpng.org/pub/png/libpng.html) library
 * [Python 3](https://www.python.org/)
 * [scipy](https://www.scipy.org/) incl. numpy headers
 * [PyQt5](https://www.riverbankcomputing.com/software/pyqt/download5) incl. QWebEngine (QWebKit also works)
-* [lapack](www.netlib.org/lapack/) library
 
 If unsure, continue with the installation and check the error messages of CMake.
 
