### 2018-11-15 #1027 @fospald
> @fospald I am getting an error "Could NOT find OpenMP_C (missing: OpenMP_C_FLAGS OpenMP_C_LIB_NAMES)", but OpenMP is not listed as a dependency.
@ctdegroot fixed that
