### 2018-11-15 #1027 @fospald
> @fospald Can you pull the test cases out of this file? It would be a lot easier for users to learn the code from the test cases if they are easier to find.
@ctdegroot Yes I can, but the test cases are pretty low level and the user should use only XML or Python for doing something. I don't think the test cases are helpful for anybody, except maybe developers.
I will try to also generate a documentation for the Python interface and a short tutorial/help for the GUI and XML files. Ok?
