### 2018-11-19 #1027 @ctdegroot
> Yes I can, but the test cases are pretty low level and the user should use only XML or Python for doing something. I don't think the test cases are helpful for anybody, except maybe developers.

@fospald Ok. This review is a bit strange for me because normally the code is made accessible to developers as well as users. I suppose that is fine if the scope is made clear.

@jedbrown Can you help clarify the position of JOSS on GUI-only type software contributions? There's not much available here to understand the code, but the GUI is well documented.
