### 2018-11-20 #1027 @jedbrown
I can solicit more specific input from the editorial board, but [the policy states](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) (among other points):

1. The software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)
2. The software should be feature complete (no half-baked solutions) and designed for maintainable extension (not one-off modifications). Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.

It doesn't really matter whether the interface is a GUI, a library, or something else, but it does matter that the software provide a good platform for research.
