### 2018-11-20 #1027 @ctdegroot
@jedbrown This package does seem to provide a platform for research. I was a bit unsure what to do because the code can’t really be understood without a lot of effort.

I’ll finish my review once the installation issues I’m having are solved.
