### 2018-11-20 #1027 @fospald
> I’ll finish my review once the installation issues I’m having are solved.

@ctdegroot can I help you with the installation issues?
