### 2018-11-20 #1027 @ctdegroot
@fospald I submitted an issue on the repository; let's discuss there. I have OpenMP installed, but CMake isn't finding it. I'm not that familiar with CMake, so if the error message gives you any clues, please let me know.
