### 2018-11-20 5222c3b
updated documentation
