+++ src/fibergen.cpp
@@ -26564,6 +26564,7 @@ public:
 		fg()->set_loadstep_callback(boost::bind(&PyFG::loadstep_callback, this));
 	}
 
+	/// Enable/Disable Python evaluation of expressions
 	void set_py_enabled(bool enabled)
 	{
 		PY::instance().set_enabled(enabled);
@@ -26859,49 +26860,49 @@ public:
 		void (PyFG::*PyFG_set_int)(const std::string& key, long value) = &PyFG::set;
 		void (PyFG::*PyFG_set)(const std::string& key) = &PyFG::set;
 
-		this->FG = py::class_<PyFG, boost::noncopyable>("FG")
-			.def("init", py::raw_function(&PyFGInit, 0))
-			.def("init_lss", &PyFG::init_lss)
-			.def("init_fibers", &PyFG::init_fibers)
-			.def("init_phase", &PyFG::init_phase)
-			.def("run", &PyFG::run)
-			.def("run", &PyFG::run_path)
-			.def("cancel", &PyFG::cancel)
-			.def("reset", &PyFG::reset)
-			.def("get_xml", &PyFG::get_xml)
-			.def("set_xml", &PyFG::set_xml)
-			.def("set_xml_precision", &PyFG::set_xml_precision)
-			.def("get_xml_precision", &PyFG::get_xml_precision)
-			.def("load_xml", &PyFG::load_xml)
-			.def("set", PyFG_set_string)
-			.def("set", PyFG_set_double)
-			.def("set", PyFG_set_int)
-			.def("set", PyFG_set)
-			.def("set", py::raw_function(&SetParameters, 1))
-			.def("get", &PyFG::get)
-			.def("erase", &PyFG::erase)
-			.def("get_phase_names", &PyFG::get_phase_names)
-			.def("get_volume_fraction", &PyFG::get_volume_fraction)
-			.def("get_real_volume_fraction", &PyFG::get_real_volume_fraction)
-			.def("get_solve_time", &PyFG::get_solve_time)
-			.def("get_distance_evals", &PyFG::get_distance_evals)
-			.def("get_residuals", &PyFG::get_residuals)
-			.def("get_effective_property", &PyFG::get_effective_property)
-			.def("get_rve_dims", &PyFG::get_rve_dims)
-			.def("get_A2", &PyFG::get_A2)
-			.def("get_A4", &PyFG::get_A4)
-			.def("get_error", &PyFG::get_error)
-			.def("get_mean_stress", &PyFG::get_mean_stress)
-			.def("get_mean_strain", &PyFG::get_mean_strain)
-			.def("get_mean_cauchy_stress", &PyFG::get_mean_cauchy_stress)
-			.def("get_mean_energy", &PyFG::get_mean_energy)
-			.def("get_field", py::raw_function(&GetField, 1))
-			.def("get_B_from_A", &PyFG::get_B_from_A)
-			.def("set_convergence_callback", &PyFG::set_convergence_callback)
-			.def("set_loadstep_callback", &PyFG::set_loadstep_callback)
-			.def("set_variable", &PyFG::set_variable)
-			.def("set_log_file", &PyFG::set_log_file)
-			.def("set_py_enabled", &PyFG::set_py_enabled)
+		this->FG = py::class_<PyFG, boost::noncopyable>("FG", "The fibergen solver class")
+			.def("init", py::raw_function(&PyFGInit, 0), "Initialize the object with a Python fibergen reference to it self, which can be used (in Python scripts) within a project file as 'fg'. If not called 'fg' will not be available. This step is necessary to expose the correctly wrapped Python object to C++, e.g. fg = fibergen.FG(); fg.init(fg)")
[…]
