### 2018-11-20 e2d568f
updated documentation
