+++ TUTORIAL.md
@@ -1,0 +1,310 @@
+
+# fibergen Tutorial
+
+## GUI
+
+The GUI is opend using the shell command
+```bash
+fibergen-gui
+```
+After startup the GUI offers to create an "empty" project, to open one of the demos from the "demos" folder
+or to open any other project using the "Open" menu. Alternatively the project filename to be opened can be appended to the command.
+
+![fibergen GUI](../screenshot_1.png "fibergen GUI")
+
+
+Creating an "empty" project shows the XML editor with the following contents
+```xml
+<?xml version="1.0" encoding="utf-8"?>
+<settings>
+	<title>Empty project</title>
+	<solver n="32">
+		<mode>elasticity</mode>
+		<materials>
+			<matrix E="1.0" nu="0.3" />
+			<fiber E="1.0" nu="0.3" />
+		</materials>
+	</solver>
+	<actions>
+		<select_material name="fiber" />
+		<place_fiber R="0.4" />
+	</actions>
+</settings>
+```
+
+![fibergen GUI](../screenshot_2.png "fibergen GUI")
+
+
+The editor shows context help for the XML element at the cursor position in the bottom window.
+There you also find a list all available settings in this context.
+
+The "empty" project template actually creates a [0,1]³ representative volume element (RVE) discretized with 32 voxels in each direction.
+Two linear isotropic materials are specified with a Young's modulus of 1.0 and a Poisson's ratio of 0.3.
+Also we specify that we want to solve an "elasticity" problem.
+
+In the actions section we then specify commands, which are run in sequence.
+First we select the current material and then place a fiber with radius 0.4 into the RVE.
+The default fiber type has spherical end caps and zero length and is placed by default in the center of the RVE.
+You can see the available settings by moving the cursor on "place_fiber".
+
+By clicking on "Run" in the menu bar the solver is run. In this case it just initializes and discretizes the geometry.
+
+![fibergen GUI](../screenshot_3.png "fibergen GUI")
+
+For each click on "Run" a new tab for the postprocessing view is opened. So it is possible to compared different runs.
+(Note: you can rename the tabs by double-clicking on them).
+
+In order to solve an elasticity problem you have to append to the actions
+```xml
+<run_load_case e11="1.0" />
+```
+which solves an homogenization problem with periodic boundary conditions and prescribed (average) elastic strain of 1.0 for the 11 (xx) component. In order to get interesting results you also have to set for example E="2.0" for the fiber material.
+
+![fibergen GUI](../screenshot_4.png "fibergen GUI")
+
+In the top row of each tab you can select the current solution view. "mid" means to show the material id. This is followed by a list of the phases (materials).
+On the right are buttons to save the data as VTK file or to write the current view as PNG image.
+The "Results" button switches to a view with a numerical summary of the solution. The "XML" button shows the XML project code that was used to produce the results.
+You can close the Results/XML view by clicking again the button.
+
+The second row of the postprocessing view is used to specify a 2d slice through the data for visualization in the bottom plot.
+In the screenshot for example we slice at z=15 and the first loadcase (we have only one loadcase since this is a linear problem and we did not explicitly specify more loadcases). But in this case we can add an loadcase by appending another
+```xml
+<run_load_case e22="1.0" />
+```
+
+The third row let's us select a color map for visualization and the contrast (reduces visualization bounds in order to show more details by cutting off peaks). With checked "depth mode" and a selected phase in the first row, allows to have a depth view of the geometry by making phase values less than 1 transparent along the slicing direction. If "interpolation" is checked bi-cubic interpolation is used for visualization of the voxel data, otherwise nearest-neighbour interpolation.
+
[…]
