### 2018-11-20 #1027 @fospald
@ctdegroot I added a tutorial and Python interface docstring to the documentation/source code.
