### 2018-11-22 0f9f2bf
fixed OpenMP
