+++ CMakeLists.txt
@@ -6,20 +6,15 @@ set(CMAKE_VERBOSE_MAKEFILE 1)
 # Note for the use of -ffast-math:
 # http://stackoverflow.com/questions/22931147/stdisinf-does-not-work-with-ffast-math-how-to-check-for-infinity
 
-SET(RELEASE_BUILD TRUE)
-#SET(RELEASE_BUILD FALSE)
+OPTION(RELEASE_BUILD "Relase build" ON)
+OPTION(USE_OPENMP "Use OpenMP" OFF)
 
 IF(RELEASE_BUILD)
-	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNAN_FILL")
 	SET(CMAKE_BUILD_TYPE Release)
-	#SET(CMAKE_BUILD_TYPE RelWithDebInfo)
-	#SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -pipe -Wall -Wunused-variable -Wno-unused-local-typedefs -Wno-long-long -ansi -march=native -O2 -malign-double -m64 -funroll-loops -fomit-frame-pointer -fopenmp -DNDEBUG")
-	#SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -march=native -O2 -fopenmp -ffast-math -fno-finite-math-only -DNDEBUG")
-	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -march=native -O2 -fopenmp -DNDEBUG")
+	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wno-unknown-pragmas -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -march=native -O2 -DNDEBUG")
 ELSE()
 	SET(CMAKE_BUILD_TYPE Debug)
-#	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DDEBUG -DNAN_FILL")
-#	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNAN_FILL")
+	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNAN_FILL")
 ENDIF()
 
 # Add cmake dir to search path
@@ -46,16 +41,31 @@ FIND_PACKAGE(PNG REQUIRED)
 INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${PNG_INCLUDE_DIR})
 
 # Find OpenMP
-FIND_PACKAGE(OpenMP REQUIRED)
-INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${LAPACK_INCLUDE_DIR})
-SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
+IF(USE_OPENMP)
+	SET(OpenMP_CXX_FLAGS "-fopenmp")
+	IF(APPLE)
+		IF(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+			SET(OpenMP_CXX "${CMAKE_CXX_COMPILER}")
[…]
