### 2018-11-22 5405f50
fixed OpenMP
