+++ src/fibergen.cpp
@@ -26920,7 +26920,7 @@ public:
 			.def("get_mean_strain", &PyFG::get_mean_strain, "Returns the mean strain of the current solution", py::args("self"))
 			.def("get_mean_cauchy_stress", &PyFG::get_mean_cauchy_stress, "Returns the mean Cauchy stress of the current solution", py::args("self"))
 			.def("get_mean_energy", &PyFG::get_mean_energy, "Returns the mean energy of the current solution", py::args("self"))
-			.def("get_field", py::raw_function(&GetField, 1), "Returns solution data as numpy ndarray. Available solution fields are: 'epsilon' (strain), 'sigma' (stress), 'u' (displacement), 'p' (pressure), 'orientation' (fiber orientation vector), 'material_id' (material id), 'fiber_id' (id of the closest fiber), 'distance' (distance to closest interface), 'phi' (phases as listed in the <materials> section), any material name.The first dimension of the returned array denotes the loadstep, the last 3 dimensions address the data.")
+			.def("get_field", py::raw_function(&GetField, 1), "Returns solution data as numpy ndarray. Available solution fields are: 'epsilon' (strain), 'sigma' (stress), 'u' (displacement), 'p' (pressure), 'orientation' (fiber orientation vector), 'material_id' (material id), 'fiber_id' (id of the closest fiber), 'distance' (distance to closest interface), 'phi' (phases as listed in the <materials> section), any material name. The first dimension of the returned array denotes the component (3 for vectors, 6 (11,22,33,23,13,12) for symmetric 3x3 matrices, 9 (11,22,33,23,13,12,32,31,21) for 3x3 matrices), the last 3 dimensions address the spatial coordinates.")
 			.def("get_B_from_A", &PyFG::get_B_from_A, "Get angular central Gaussian covariance matrix from moment matrix", py::args("self", "A"))
 			.def("set_convergence_callback", &PyFG::set_convergence_callback, "Set a callback function to be called each iteration of the solver. If the callback returns True, the solver is canceled.", py::args("self", "func"))
 			.def("set_loadstep_callback", &PyFG::set_loadstep_callback, "Set a callback function to be called each loadstep of the solver. If the callback returns True, the solver is canceled.", py::args("self", "func"))
