### 2018-11-23 fospald/fibergen#6 @ctdegroot
Finding boost libraries on MacOS
I am getting the following error on MacOS

```
CMake Error at /usr/local/Cellar/cmake/3.12.4/share/cmake/Modules/FindBoost.cmake:2048 (message):
  Unable to find the requested Boost libraries.

  Boost version: 1.68.0

  Boost include path: /usr/local/include

  Could not find the following Boost libraries:

          boost_thread (missing dependency: Threads)
          boost_python-py37

  Some (but not all) of the required Boost libraries were found.  You may
  need to install these additional Boost libraries.  Alternatively, set
  BOOST_LIBRARYDIR to the directory containing Boost libraries or BOOST_ROOT
  to the location of Boost.
```

My main boost libraries are in `/usr/local/Cellar/boost/1.68.0/lib/`, but the boost-python library is in `/usr/local/Cellar/boost-python3/1.68.0/lib/` and is called `libboost_python37.dylib`.

I tried adding this to CMakeLists.txt:

```
SET(PYTHON_VERSION_MAJOR 3)
SET(PYTHON_VERSION_MINOR 7)
```
