### 2018-11-23 cc50337
updated build script
