+++ README.md
@@ -25,7 +25,7 @@ A FFT-based homogenization tool.
 The following libraries are required, which are likely already installed on your system:
 * [CMake](https://cmake.org/)
 * [gcc](https://gcc.gnu.org/)
-* [OpenMP](https://www.openmp.org/)
+* [OpenMP](https://www.openmp.org/) for parallelization (optional)
 * [boost](https://www.boost.org/) incl. boost.python 3 and [boost-numeric-bindings](https://mathema.tician.de/software/boost-numeric-bindings/)
 * [FFTW3](http://www.fftw.org/) library
 * [lapack](www.netlib.org/lapack/) library
