+++ CMakeLists.txt
@@ -9,6 +9,10 @@ set(CMAKE_VERBOSE_MAKEFILE 1)
 OPTION(RELEASE_BUILD "Relase build" ON)
 OPTION(USE_OPENMP "Use OpenMP" ON)
 
+IF(APPLE)
+	set(CMAKE_SHARED_LIBRARY_SUFFIX ".so")
+ENDIF()
+
 IF(RELEASE_BUILD)
 	SET(CMAKE_BUILD_TYPE Release)
 	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wno-unknown-pragmas -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -march=native -O2 -DNDEBUG")
@@ -71,10 +75,17 @@ ENDIF()
 #MESSAGE("ITK include dirs: " ${ITK_INCLUDE_DIRS})
 #MESSAGE("ITK libraries: " ${ITK_LIBRARY_DIRS})
 
+# Fint threads library https://cmake.org/cmake/help/v3.2/module/FindThreads.html
+FIND_PACKAGE(Threads REQUIRED)
+MESSAGE(${CMAKE_THREAD_LIBS_INIT})
+
 # Find Boost library
 SET(Boost_USE_MULTITHREAD ON)
-SET(Boost_Python python)
-SET(Boost_Python python-py${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
+IF(APPLE)
+	SET(Boost_Python python${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
+ELSE()
+	SET(Boost_Python python-py${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
+ENDIF()
 FIND_PACKAGE(Boost REQUIRED COMPONENTS system program_options thread ${Boost_Python} iostreams filesystem)
 INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${Boost_INCLUDE_DIRS})
 LINK_DIRECTORIES(${LINK_DIRECTORIES} ${Boost_LIBRARY_DIRS})
@@ -103,7 +114,7 @@ ENDIF()
 
 SET(SOURCES src/fibergen.cpp)
 
-SET(LIBS ${Boost_LIBRARIES} ${PNG_LIBRARY} ${FFTWF_LIB} ${FFTWD_LIB} ${FFTWL_LIB} ${LAPACK_LIBRARIES} ${PYTHON_LIBRARIES} ${ITK_LIBRARIES} pthread)
+SET(LIBS ${Boost_LIBRARIES} ${PNG_LIBRARY} ${FFTWF_LIB} ${FFTWD_LIB} ${FFTWL_LIB} ${LAPACK_LIBRARIES} ${PYTHON_LIBRARIES} ${ITK_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT})
 SET(PROJECT_BINARY_DIR ${CMAKE_SOURCE_DIR}/bin)
 
 SITE_NAME(HOST_NAME)
