### 2018-11-23 fospald/fibergen#6 @fospald
@ctdegroot I updated the CMakeLists.txt
I hope it will work. You have to set PYTHON_VERSION_MAJOR and PYTHON_VERSION_MINOR again.
