### 2018-11-23 9c04a6a
fixed CMakeLists.txt
