+++ CMakeLists.txt
@@ -7,15 +7,40 @@ set(CMAKE_VERBOSE_MAKEFILE 1)
 # http://stackoverflow.com/questions/22931147/stdisinf-does-not-work-with-ffast-math-how-to-check-for-infinity
 
 OPTION(RELEASE_BUILD "Relase build" ON)
+OPTION(DEBUG_INFO_BUILD "Enable debug info on release build" OFF)
 OPTION(USE_OPENMP "Use OpenMP" ON)
 
 IF(APPLE)
-	set(CMAKE_SHARED_LIBRARY_SUFFIX ".so")
+	SET(CMAKE_SHARED_LIBRARY_SUFFIX ".so")
+ENDIF()
+
+# Compiler check
+IF(CMAKE_CXX_COMPILER_ID MATCHES "GNU")
+	SET(GNU_COMPILER ON)
+ELSE()
+	SET(GNU_COMPILER OFF)
+ENDIF()
+IF(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+	SET(CLANG_COMPILER ON)
+ELSE()
+	SET(CLANG_COMPILER OFF)
 ENDIF()
 
 IF(RELEASE_BUILD)
-	SET(CMAKE_BUILD_TYPE Release)
-	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wno-unknown-pragmas -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -march=native -O2 -DNDEBUG")
+	IF(DEBUG_INFO_BUILD)
+		SET(CMAKE_BUILD_TYPE ReleaseWithDebInfo)
+		IF(GNU_COMPILER)
+			SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O2 -g")
+		ENDIF()
+	ELSE()
+		SET(CMAKE_BUILD_TYPE Release)
+		IF(GNU_COMPILER)
+			SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3")
+		ENDIF()
+	ENDIF()
+	IF(GNU_COMPILER)
+		SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -march=native -DNDBUG -Wall -Wno-unknown-pragmas -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long")
+	ENDIF()
 ELSE()
 	SET(CMAKE_BUILD_TYPE Debug)
 	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNAN_FILL")
@@ -48,7 +73,7 @@ INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${PNG_INCLUDE_DIR})
 IF(USE_OPENMP)
 	SET(OpenMP_CXX_FLAGS "-fopenmp")
 	IF(APPLE)
-		IF(CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+		IF(CLANG_COMPILER)
 			SET(OpenMP_CXX "${CMAKE_CXX_COMPILER}")
 			SET(OpenMP_CXX_FLAGS "-fopenmp=libomp -Wno-unused-command-line-argument")
 			SET(OpenMP_CXX_LIB_NAMES "libomp" "libgomp" "libiomp5")
@@ -71,13 +96,11 @@ LINK_DIRECTORIES(${LINK_DIRECTORIES} ${ITK_LIBRARY_DIRS})
 IF(ITK_FOUND)
 	ADD_DEFINITIONS(-DITK_ENABLED)
 ENDIF()
-
[…]
