### 2018-11-23 edf39bf
extended Python functionality
