+++ lib/fibergen_gui.py
@@ -10,6 +10,7 @@ import webbrowser
 import base64
 import copy
 import cgi
+import pydoc
 import traceback
 import codecs
 import collections
@@ -1575,23 +1576,31 @@ class XMLHighlighter(QtGui.QSyntaxHighlighter):
 		commentFormat.setForeground(QtCore.Qt.gray)
 		self.highlightingRulesPython.append((QtCore.QRegExp("#.*"), commentFormat, 0))
 
+		self.pythonOnly = False
+		if self.pythonOnly:
+			self.pythonStartExpression = QtCore.QRegExp("^")
+			self.pythonEndExpression = QtCore.QRegExp("$")
+
+
 	def highlightBlock(self, text):
 		
-		#for every pattern
-		for pattern, format in self.highlightingRules:
+		if not self.pythonOnly:
 
-			#Check what index that expression occurs at with the ENTIRE text
-			index = pattern.indexIn(text)
+			#for every pattern
+			for pattern, format in self.highlightingRules:
 
-			#While the index is greater than 0
-			while index >= 0:
+				#Check what index that expression occurs at with the ENTIRE text
+				index = pattern.indexIn(text)
 
-				#Get the length of how long the expression is true, set the format from the start to the length with the text format
-				length = pattern.matchedLength()
-				self.setFormat(index, length, format)
+				#While the index is greater than 0
+				while index >= 0:
 
-				#Set index to where the expression ends in the text
-				index = pattern.indexIn(text, index + length)
+					#Get the length of how long the expression is true, set the format from the start to the length with the text format
+					length = pattern.matchedLength()
+					self.setFormat(index, length, format)
+
+					#Set index to where the expression ends in the text
+					index = pattern.indexIn(text, index + length)
 
 		Flag_Comment = 1
 		Flag_Python = 2
@@ -1990,6 +1999,31 @@ class HelpWidget(QtWebKitWidgets.QWebView):
 
 		self.updateHelpPath(items, inside)
 
+	def getCursorHelp(self):
+
+		c = self.editor.textCursor()
+		pos = c.position()
+		txt = self.editor.toPlainText()
+
+		for m in re.finditer(r'\b\w+\b', txt):
+			if pos >= m.start() and pos <= (m.start() + len(m.group(0))):
+				word = m.group(0)
+				if word == 'fg':
+					word = 'fibergen'
+				for k in [word, "fibergen.%s" % word, "fibergen.FG.%s" % word]:
+					try:
+						#helpstr = pydoc.render_doc(k, "Help on %s", renderer=pydoc.plaintext)
+						#helpstr = '<pre>' + cgi.escape(helpstr) + '</pre>'
+						helpstr = pydoc.render_doc(k, "Help on %s", renderer=pydoc.html)
+						helpstr = helpstr.replace('&nbsp;', ' ')
+						return helpstr
+					except:
+						pass
+				break
+
+		helpstr = "Unknown element"
+		return helpstr
+
 	def updateHelpPath(self, items, inside=False):
 
 		#if len(items):
@@ -2045,11 +2079,11 @@ p ~ p {
 			return '<a href="http://x#help#' + '#'.join(apath) + '">' + tag + "</a>"
 		
 		if en is None:
-			help = "Unknown element"
+			helpstr = self.getCursorHelp()
 		else:
-			help = cgi.escape(e.get("help"))
+			helpstr = cgi.escape(e.get("help"))
 
-		html += '<div class="help">' + help + "</div>"
+		html += '<div class="help">' + helpstr + "</div>"
 
[…]
