### 2018-11-25 fospald/fibergen#6 @ctdegroot
This issue seems to be solved now.
