### 2018-11-25 fospald/fibergen#7 @ctdegroot
ITK Dependencies
When installing I am getting errors about ITK, which is not listed as a dependency in the README. The error message is below. Is this an actual dependency or something that can be made optional?

```
  By not providing "FindITK.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "ITK", but
  CMake did not find one.

  Could not find a package configuration file provided by "ITK" with any of
  the following names:

    ITKConfig.cmake
    itk-config.cmake

  Add the installation prefix of "ITK" to CMAKE_PREFIX_PATH or set "ITK_DIR"
  to a directory containing one of the above files.  If "ITK" provides a
  separate development package or SDK, be sure it has been installed.
```
