### 2018-11-26 #1027 @fospald
@ctdegroot ITK is optional. There must be another error message if it does not compile.
