### 2018-11-26 c65b90c
updated CMakeLists.txt
