+++ CMakeLists.txt
@@ -6,6 +6,9 @@ set(CMAKE_VERBOSE_MAKEFILE 1)
 # Note for the use of -ffast-math:
 # http://stackoverflow.com/questions/22931147/stdisinf-does-not-work-with-ffast-math-how-to-check-for-infinity
 
+SET(PYTHON_VERSION_MAJOR 3)
+SET(PYTHON_VERSION_MINOR 5)
+
 OPTION(RELEASE_BUILD "Relase build" ON)
 OPTION(DEBUG_INFO_BUILD "Enable debug info on release build" OFF)
 OPTION(USE_OPENMP "Use OpenMP" ON)
@@ -50,8 +53,6 @@ ENDIF()
 LIST(APPEND CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/cmake")
 
 # Find Python library and executable
-SET(PYTHON_VERSION_MAJOR 3)
-SET(PYTHON_VERSION_MINOR 5)
 FIND_PACKAGE(PythonInterp ${PYTHON_VERSION_MAJOR} REQUIRED)
 FIND_PACKAGE(PythonLibs ${PYTHON_VERSION_MAJOR} REQUIRED)
 MESSAGE("Python executable: " ${PYTHON_EXECUTABLE})
@@ -82,7 +83,7 @@ IF(USE_OPENMP)
 			SET(OpenMP_libiomp5_LIBRARY ${OpenMP_CXX_LIB_NAMES})
 		ENDIF()
 	ENDIF()
-	FIND_PACKAGE(OpenMP)
+	FIND_PACKAGE(OpenMP QUIET)
 	IF(OpenMP_FOUND)
 		SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
 		ADD_DEFINITIONS(-DOPENMP_ENABLED)
@@ -90,7 +91,7 @@ IF(USE_OPENMP)
 ENDIF()
 
 # Find ITK library
-FIND_PACKAGE(ITK)
+FIND_PACKAGE(ITK QUIET)
 INCLUDE_DIRECTORIES(${INCLUDE_DIRECTORIES} ${ITK_INCLUDE_DIRS})
 LINK_DIRECTORIES(${LINK_DIRECTORIES} ${ITK_LIBRARY_DIRS})
 IF(ITK_FOUND)
