### 2018-11-26 fospald/fibergen#7 @fospald
The warning will no longer be displayed if ITK is not found.
