### 2018-11-30 #1027 @fospald
@jedbrown @ctdegroot How to continue here?
