### 2018-11-30 #1027 @ctdegroot
@fospald The ITK issue is solved but there is another package that needs to be installed on my system and my package manager doesn’t include it. I’ll have to install manually. I haven’t really had a chance to work on it this week (end of teaching semester) but I will get back to it as soon as I can.
