### 2018-11-30 #1027 @fospald
@ctdegroot I can also just say that MacOS is not supported yet, if it is too time consuming.
