### 2018-11-30 #1027 @ctdegroot
@fospald I’ll try on Linux so that I can finish my review sooner
