### 2018-11-30 #1027 @ctdegroot
@fospald I can definitely help work out the issues on MacOS, since I bet this will be beneficial for your users. I'll do Linux as well so that I can at least verify everything works there first.
