### 2018-12-06 #1027 @ctdegroot
@fospald Seems like I have fixed all of the remaining cmake dependency issues. The last one was to install boost-numeric-bindings. I am now having the following error:

```
fatal error: 'boost/gil/extension/io/png_dynamic_io.hpp' file not found
```

Any idea where this comes from?

Also, I should note that I had to set the following in the file `CMakeLists.txt`:

``
set (CMAKE_CXX_STANDARD 11)
```

Otherwise there were fatal errors due to templates not having a space between the angled brackets, i.e. (">>" vs "> >"). Seems the clang compiler is picky about this.
