### 2018-12-06 #1027 @ctdegroot
@fospald I also noted that there were errors because my host name had periods in it, so I added the following to `CMakeLists.txt`:

```
STRING(REPLACE "." "_" HOST_NAME ${HOST_NAME})
```
