### 2018-12-07 #1027 @fospald
> ```
> fatal error: 'boost/gil/extension/io/png_dynamic_io.hpp' file not found
> ```
>
> Any idea where this comes from?
@ctdegroot Looks like boost gill is missing an extension. Do you have libpng installend? Maybe there is some boost-gil-png-dev extension package? I will try to avoid inclusion of this header.
