+++ build.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-HOST=$(uname -n | tr " -" "__")
+HOST=$(uname -n | tr " -." "___")
 FIBERGEN_DIR=$(cd $(dirname "$0") && pwd)
 
 BUILD_DIR="$FIBERGEN_DIR/build/$HOST"
