### 2018-12-07 #1027 @fospald
@ctdegroot Looks like since Boost 1.68, Boost.GIL is a separate module.
https://github.com/boostorg
Not sure if this solves the issue, but then you maybe have to install this module separately with your installer. I'll update the CMake file.
