### 2018-12-07 be7295f
fixed CMakeLists.txt
