### 2018-12-07 a260eb7
updated CMakeLists.txt
