+++ lib/fibergen_gui.py
@@ -21,6 +21,7 @@ import numpy as np
 import scipy.misc
 import keyword
 import textwrap
+import signal
 from weakref import WeakKeyDictionary
 
 from PyQt5 import QtCore, QtGui, QtWidgets
@@ -739,6 +740,28 @@ class PlotWidget(QtWidgets.QWidget):
 		hbox.addWidget(self.replotButton)
 		vbox.addLayout(hbox)
 
+		def makeMaskFieldCallback(index):
+			return lambda checked: self.maskField(index, checked)
+
+		materials = field_groups[1]
+		if len(materials) >= 2:
+			#hbox = QtWidgets.QHBoxLayout()
+			hbox.addWidget(QtWidgets.QLabel("Mask:"))
+			gbox = QtWidgets.QHBoxLayout()
+			for i, field in enumerate(materials):
+				button = QtWidgets.QToolButton()
+				button.setText(field.label)
+				if len(field.description):
+					button.setToolTip(field.description)
+				button.setCheckable(True)
+				button.toggled.connect(makeMaskFieldCallback(index))
+				if i > 0:
+					gbox.addSpacing(spacing)
+				gbox.addWidget(button)
+				field.mask_button = button
+			hbox.addLayout(gbox)
+			#vbox.addLayout(hbox)
+
 		self.stack = QtWidgets.QStackedWidget()
 		self.stack.setFrameShape(QtWidgets.QFrame.StyledPanel)
 		self.stack.setFrameShadow(QtWidgets.QFrame.Sunken)
@@ -880,9 +903,10 @@ class PlotWidget(QtWidgets.QWidget):
 		except:
 			print(traceback.format_exc())
 
-		self.resetBounds()
+		data = self.getCurrentSlice()
+		self.resetBounds(data)
 		self.replotSuspended = False
-		self.replot()
+		self.replot(data)
 
 		self.updateFigCanvasVisible()
 
@@ -1189,8 +1213,9 @@ class PlotWidget(QtWidgets.QWidget):
 		self.replot()
 	
 	def loadstepSliderChanged(self):
-		self.resetBounds()
-		self.replot()
+		data = self.getCurrentSlice()
+		self.resetBounds(data)
+		self.replot(data)
 		self.loadstepLabel.setText("%04d" % self.loadstepSlider.value())
 
 	def sliceComboChanged(self, index):
@@ -1215,9 +1240,10 @@ class PlotWidget(QtWidgets.QWidget):
 		enable = (state != 0)
 		self.vminText.setEnabled(enable)
 		self.vmaxText.setEnabled(enable)
-		self.resetBounds()
+		data = self.getCurrentSlice()
+		self.resetBounds(data)
 		if (state == 0):
-			self.replot()
+			self.replot(data)
 
 	def setAlpha(self, a):
 		a_max = 0.4999
@@ -1230,6 +1256,10 @@ class PlotWidget(QtWidgets.QWidget):
 	def getCurrentSlice(self):
 		if self.currentFieldIndex is None:
 			return None
+
+		#print("getCurrentSlice")
+		#traceback.print_stack()
+
 		field = self.fields[self.currentFieldIndex]
 		s_index = self.sliceSlider.value()
 		ls_index = self.loadstepSlider.value()
@@ -1242,27 +1272,47 @@ class PlotWidget(QtWidgets.QWidget):
 
 		s_index_end = s_index + min(depth, depth_max)
 
+		# get mask
+		mask_fields = []
+		materials = self.field_groups[1]
+		if len(materials) >= 2:
+			for i, mfield in enumerate(materials):
+				if mfield.mask_button.isChecked():
+					mask_fields.append(mfield)
+
 		if (sliceIndex == 0):
 			data = field.data[ls_index][field.component,s_index:s_index_end,:,:]
[…]
