+++ CMakeLists.txt
@@ -50,6 +50,10 @@ IF(RELEASE_BUILD)
 ELSE()
 	SET(CMAKE_BUILD_TYPE Debug)
 	SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DNAN_FILL")
+	SET(Boost_USE_DEBUG_PYTHON ON)
+	SET(Boost_USE_DEBUG_RUNTIME ON)
+	SET(Boost_USE_RELEASE_LIBS OFF)
+	SET(Boost_USE_DEBUG_LIBS ON)
 ENDIF()
 
 # Add cmake dir to search path
@@ -107,7 +111,8 @@ ENDIF()
 FIND_PACKAGE(Threads REQUIRED)
 
 # Find Boost library
-SET(Boost_USE_MULTITHREAD ON)
+SET(Boost_USE_MULTITHREADED ${USE_OPENMP})
+SET(Boost_USE_STATIC_RUNTIME OFF)
 IF(APPLE)
 	SET(Boost_Python python${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
 ELSE()
