### 2018-12-07 #1027 @ctdegroot
@fospald Ok thanks for the updates. I do have libpng and zlib installed. I will look into installing Boost.GIL separately and hopefully that solves it.
