### 2018-12-07 #1027 @ctdegroot
@fospald I've just installed the headers for boost.gil but as you can see from the repo that you linked, the png_dynamic_io.hpp and png_io.hpp headers are not present in that version.
