### 2018-12-10 #1027 @fospald
@ctdegroot In boost 1.68 they changed the GIL headers completely. I tried to fix this now.
If there is still an error let me know. It is probably line 7054 of fibergen.cpp. You can temporarily disable this line in this case until I have a fix.
