+++ README.md
@@ -1,6 +1,7 @@
 <p align="center">
   <a href="LICENSE" alt="GPLv3 license"><img src="https://img.shields.io/badge/license-GPLv3-brightgreen.svg" /></a>
   <a href="#" alt="no warranty"><img src="https://img.shields.io/badge/warranty-no-red.svg" /></a>
+  <a href="http://joss.theoj.org/papers/add4f02dcedfaae8f81da200000d64ee"><img src="http://joss.theoj.org/papers/add4f02dcedfaae8f81da200000d64ee/status.svg"></a>
 </p>
 
 # fibergen
