### 2018-12-29 #1027 @jedbrown
Testing on Linux with boost-1.68 and python-3.7, I need the following (i.e., suggest eliminating the switch unless it's needed for some reason).
```
diff --git i/CMakeLists.txt w/CMakeLists.txt
index 665655d..bad9ce8 100644
--- i/CMakeLists.txt
+++ w/CMakeLists.txt
@@ -116,7 +116,7 @@ SET(Boost_USE_STATIC_RUNTIME OFF)
 IF(APPLE)
 	SET(Boost_Python python${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
 ELSE()
-	SET(Boost_Python python-py${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
+	SET(Boost_Python python${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
 ENDIF()
 # TODO: Boost >= 1.68: also add gil component?
 FIND_PACKAGE(Boost REQUIRED COMPONENTS system program_options thread ${Boost_Python} iostreams filesystem)
```

My build still fails (gcc-8.2.1).
```
/bin/c++  -DFFTWF_ENABLED -DFG_PYTHON_HEADER_NAME=\"/home/jed/joss/fibergen/build/joule/PythonModuleHeader_joule.h\" -DHOST_NAME=joule -DLIB_BUILD -DOPENMP_ENABLED -I/usr/include/python3.7m -
I/usr/lib/python3.7/site-packages/numpy/core/include  -O3 -march=native -DNDBUG -Wall -Wno-unknown-pragmas -Wno-unused-variable -Wno-unused-local-typedefs -Wno-long-long -fopenmp -O3 -DNDEBUG
   -std=gnu++11 -o CMakeFiles/fibergen_joule.dir/src/fibergen.cpp.o -c /home/jed/joss/fibergen/src/fibergen.cpp
/home/jed/joss/fibergen/src/fibergen.cpp: In member function ‘void FiberGenerator<T, DIM>::writeDistanceMap(const string&, const boost::numeric::ublas::c_vector<T, DIM>&, const boost::numeric
::ublas::c_vector<T, DIM>&, const boost::numeric::ublas::c_vector<T, DIM>&, std::size_t, std::size_t, T, T, T, bool, int) const’:
/home/jed/joss/fibergen/src/fibergen.cpp:7054:49: error: no matching function for call to ‘write_view(const string&, const const_view_t)’
   gil::write_view(filename, gil::const_view(img));
                                                 ^
In file included from /usr/include/boost/gil/extension/io/png/write.hpp:29,
                 from /usr/include/boost/gil/extension/io/png.hpp:23,
                 from /home/jed/joss/fibergen/src/fibergen.cpp:150:
/usr/include/boost/gil/io/write_view.hpp:40:6: note: candidate: ‘template<class Writer, class View> void boost::gil::write_view(Writer&, const View&, typename boost::enable_if<typename boost$
:mpl::and_<typename boost::gil::detail::is_writer<Writer>::type, typename boost::gil::is_format_tag<typename Reader::format_tag_t>::type, typename boost::gil::is_write_supported<typename boo$
t::gil::get_pixel_type<View>::type, typename Writer::format_tag_t>::type>::type>::type*)’
 void write_view( Writer&     writer
      ^~~~~~~~~~
/usr/include/boost/gil/io/write_view.hpp:40:6: note:   template argument deduction/substitution failed:
/usr/include/boost/gil/io/write_view.hpp: In substitution of ‘template<class Writer, class View> void boost::gil::write_view(Writer&, const View&, typename boost::enable_if<typename boost::m$
l::and_<typename boost::gil::detail::is_writer<Writer>::type, typename boost::gil::is_format_tag<typename Reader::format_tag_t>::type, typename boost::gil::is_write_supported<typename boost:$
gil::get_pixel_type<View>::type, typename Writer::format_tag_t>::type>::type>::type*) [with Writer = const std::__cxx11::basic_string<char>; View = boost::gil::image_view<boost::gil::memory_$
ased_2d_locator<boost::gil::memory_based_step_iterator<const boost::gil::pixel<unsigned char, boost::gil::layout<boost::mpl::vector1<boost::gil::gray_color_t> > >*> > >]’:
/home/jed/joss/fibergen/src/fibergen.cpp:7054:49:   required from here
```
(several screens of template-related errors follow). Commenting that line allows the build to complete.
