### 2018-12-29 #1027 @ctdegroot
@fospald Sorry for the delay since your latest update. The code does compile with line 7054 commented out.
