### 2018-12-29 #1027 @ctdegroot
@fospald When I run `fibergen-gui`, I get the following error:

```
Traceback (most recent call last):
  File "/Users/ctdegroot/GitRepos/fibregen/bin/../gui/main.py", line 4, in <module>
    import fibergen_gui as fg_gui
  File "/Users/ctdegroot/GitRepos/fibregen/lib/fibergen_gui.py", line 7, in <module>
    import fibergen
  File "/Users/ctdegroot/GitRepos/fibregen/lib/fibergen.py", line 12, in <module>
    fibergen = __import__(libname)
ModuleNotFoundError: No module named 'fibergen_Christopher_DeGroots_Macbook_Pro'
```
The contents of `lib` includes the shared object file `fibergen_Christopher_DeGroots_Macbook_Pro_local.so`. Seems to be a naming issue here.
