### 2019-01-07 b930e6d
fixed build script
