+++ lib/fibergen.py
@@ -1,11 +1,12 @@
 
-import socket, os
+import socket, os, subprocess
 
 #os.system("fibergen-build")
 
-hostname = socket.gethostname()
-hostname = hostname.replace("-", "_")
-hostname = hostname.replace(" ", "_")
+hostname = subprocess.check_output(['uname', '-n']).decode("utf-8").strip()
+#hostname = socket.gethostname()
+for c in ["-", ".", " "]:
+	hostname = hostname.replace(c, "_")
 
 libname = "fibergen_" + hostname
 
@@ -14,5 +15,4 @@ locals().update(fibergen.__dict__)
 
 #fibergen_gui = __import__("fibergen_gui")
 #locals().update({'gui': fibergen_gui.__dict__})
-
 
