### 2019-01-07 #1027 @fospald
@ctdegroot @jedbrown I think I fixed the build and naming issues mentioned above. Can you please retry?
