### 2019-01-08 #1027 @ctdegroot
@fospald I get the following:

```
/Users/ctdegroot/GitRepos/fibregen/bin/fibergen-gui: line 13: 11684 Segmentation fault: 11  $PYTHON "${FIBERGEN_DIR}/../gui/main.py" "$@"
```

I should mention that Mac doesn't have `ldd` so I replaced with `otool -L` which should be equivalent.
