### 2019-01-09 #1027 @fospald
@ctdegroot
Can you run:
`fibergen demo/elasticity/hashin/project.xml`
`python3 -c "import fibergen; fg = fibergen.FG(); fg.run()"`
?
