### 2019-01-10 #1027 @ctdegroot
@fospald First command results in:

```
dyld: Library not loaded: @rpath/libpython3.7m.dylib
  Referenced from: /Users/ctdegroot/GitRepos/fibregen/bin/fibergen_cdg_mac01_eng_uwo_ca
  Reason: image not found
/Users/ctdegroot/GitRepos/fibregen/bin/fibergen: line 6: 18485 Abort trap: 6           "${FIBERGEN_DIR}/fibergen_${HOST}" "$@"
```

Second command results in:

```
Segmentation fault: 11
```

Let me know if you want to keep trying on Mac. I can finish the review on Linux if you want.
