### 2019-01-11 #1027 @fospald
@ctdegroot This seems to be an installation/search path issue for libpython3.
https://github.com/pre-commit/pre-commit/issues/618
You can finish the review on Linux if you want.
