### 2019-01-18 #1027 @fospald
@ctdegroot there is only one item left in the checklist. Does it work on Linux?
