### 2019-01-18 #1027 @ctdegroot
@fospald Still not getting through the install on Linux. First there was a problem with naming of the boost.python library. Had to make the following change in CMakeLists.txt

```
 # Try to find boost python
-SET(Boost_Python python${PYTHON_VERSION_MAJOR}${PYTHON_VERSION_MINOR})
+SET(Boost_Python python${PYTHON_VERSION_MAJOR})
```
I now have an error:

```
  Could not find the following Boost libraries:

          boost_iostreams
```

Still trying to figure that one out.
