### 2019-01-18 #1027 @ctdegroot
@fospald I am using boost version 1.60.0. I could find the library file in my installation. I've tried removing iostreams from the CMakeLists.txt file, however the build fails. It seems like this library is required. Looks like iostreams doesn't build unless libbz2 is installed first.

I have now managed to get it build but cannot import QtWebKitWidgets. I guess this is some component of PyQt5 that I don't have. Is there any way to automate the install of all these dependencies? Very tedious to install everything manually.
