### 2019-01-29 #1027 @fospald
@ctdegroot I updated the code. Can you try running with
fibergen-gui --disable-browser
This will disable all QTWebKit components. Then you have to open the demo files manually.
