### 2019-01-29 #1027 @ctdegroot
@fospald @jedbrown The software is working with the browser disabled. As such I was able to complete my review, and everything seems to be working fine. I would suggest that the installation procedure (mainly handling of dependencies) could be streamlined; however I give that as a future goal and won't let that hold up this review any further. I would suggest placing a note in the installation instructions that Mac is not fully supported, since we could not get this working.
