### 2019-01-30 #1027 @fospald
@ctdegroot Thank you for testing. I updated the README with some more installation notes.
