+++ lib/fibergen_gui.py
@@ -431,13 +431,14 @@ class MyWebPage(QtWebKitWidgets.QWebPage):
 	def setUrlFrame(self, url):
 		self.currentFrame().setUrl(url)
 
-	def defaultCSS(self, tags=True):
 
-		app = QtWidgets.QApplication.instance()
-		pal = app.palette()
-		font = QtGui.QFontDatabase.systemFont(QtGui.QFontDatabase.GeneralFont)
+def defaultCSS(tags=True):
 
-		html = """
+	app = QtWidgets.QApplication.instance()
+	pal = app.palette()
+	font = QtGui.QFontDatabase.systemFont(QtGui.QFontDatabase.GeneralFont)
+
+	html = """
 body, table {
 	font-size: """ + str(font.pointSize()) + """pt;
 }
@@ -496,10 +497,10 @@ h3 {
 	margin-top: 1.0em;
 }
 """
-		if tags:
-			html = "<style>" + html + "</style>"
+	if tags:
+		html = "<style>" + html + "</style>"
 
-		return html
+	return html
 
 
 class PlotField(object):
@@ -872,7 +873,7 @@ class PlotWidget(QtWidgets.QWidget):
 		if not app.pargs.disable_browser:
 			self.resultTextEdit = QtWebKitWidgets.QWebView()
 			self.resultPage = MyWebPage()
-			self.resultPage.setHtml(self.resultPage.defaultCSS() + resultText)
+			self.resultPage.setHtml(defaultCSS() + resultText)
 			self.resultTextEdit.setPage(self.resultPage)
 		else:
 			self.resultTextEdit = QtWidgets.QTextEdit()
@@ -1914,31 +1915,26 @@ class XMLTextEdit(QtWidgets.QTextEdit):
 		return True
 
 
-class HelpWidget(QtWebKitWidgets.QWebView):
+class HelpWidgetCommon(QtCore.QObject):
 
-	def __init__(self, editor, parent = None):
+	updateHtml = QtCore.pyqtSignal('QString')
 
-		QtWebKitWidgets.QWebView.__init__(self, parent)
+	def __init__(self, editor):
+
+		super(QtCore.QObject, self).__init__(editor)
 
 		self.editor = editor
 		self.editor.selectionChanged.connect(self.editorSelectionChanged)
 		self.editor.textChanged.connect(self.editorSelectionChanged)
 		self.editor.cursorPositionChanged.connect(self.editorSelectionChanged)
 
-		self.timer = QtCore.QTimer(self)
+		self.timer = QtCore.QTimer(editor)
 		self.timer.setInterval(100)
 		self.timer.timeout.connect(self.updateHelp)
 
 		cdir = os.path.dirname(os.path.abspath(__file__))
 
 		self.ff = ET.parse(cdir + "/../doc/fileformat.xml")
-		
-		self.mypage = MyWebPage()
-		self.mypage.linkClicked.connect(self.linkClicked)
-		self.setPage(self.mypage)
-
-		#self.setStyleSheet("background:transparent");
-		#self.setAttribute(QtCore.Qt.WA_TranslucentBackground);
 
 	def linkClicked(self, url):
 
@@ -2120,7 +2116,7 @@ class HelpWidget(QtWebKitWidgets.QWebView):
 		typ = e.get("type")
 		values = e.get("values")
 
-		html = self.mypage.defaultCSS() + """
+		html = defaultCSS() + """
 <style>
 h2 {
 	margin-top: 0;
@@ -2247,15 +2243,47 @@ p ~ p {
 				html += tags
 				html += "</table>"
 
-		# ui.textEdit->verticalScrollBar()->setValue(0);
-		self.mypage.setHtml(html)
[…]
