### 2019-02-01 #1027 @fospald
@ctdegroot @jedbrown Now I also added a fallback browser, when using "fibergen-gui --disable-browser", which at least provides the same functionality as the QtWebEngine browser but is a bit ugly.
