### 2019-02-04 #1027 @jedbrown
@fospald Can you fix the citations that use inline style to use appropriate syntax `the method of @bibkey` versus `the method of [@bibkey]`? See https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax

Looks good to me otherwise.
