+++ joss/paper.bib
@@ -44,18 +44,6 @@
   language={English},
 	file	= {555:Kabel_Böhlke_Schneider__Efficient_fixed_point_and_Newton-Krylov_solvers_for_FFT-based_homogenization_of_elasticity_at_large_deformations.pdf:PDF}
 }
-@ARTICLE{KabelMerkertSchneider2014,
-  title = {{FFT-based homogenization using composite voxels}},
-  journal = {Computer Methods in Applied Mechanics and Engineering},
-  volume = {},
-  number = {},
-  pages = {},
-  year = {2014},
-  note = {},
-  author = {M. Kabel and D. Merkert and M. Schneider},
-  doi={},
-	file	= {555:Kabel_Merkert_Schneider__FFT-based_homogenization_using_composite_voxels.pdf:PDF}
-}
 @article{KabelMerkertSchneider2015,
 	title = "Use of composite voxels in FFT-based homogenization",
 	journal = "Computer Methods in Applied Mechanics and Engineering",
@@ -106,25 +94,30 @@
   pages = {673--674},
   year = {2014}
 }
-@ARTICLE{SchneiderOspaldKabel2015,
-  author = {Schneider, Matti and Ospald, Felix and Kabel, Matthias},
-  title = {Hyperelastic laminates, composite voxels and FFT-based homogenization at finite strains},
-  journal = {International Journal for Numerical Methods in Engineering},
-  issn = {},
-  doi = {},
-  keywords = {Composite materials, Homogenization, Hyperelasticity, Laminates, Lippmann-Schwinger equation, FFT},
-  year = {2015},
-  note = {}
+@article{SchneiderOspaldKabel2015,
+	title = "A model order reduction method for computational homogenization at finite strains on regular grids using hyperelastic laminates to approximate interfaces",
+	journal = "Computer Methods in Applied Mechanics and Engineering",
+	volume = "309",
+	pages = "476 - 496",
+	year = "2016",
+	issn = "0045-7825",
+	doi = "https://doi.org/10.1016/j.cma.2016.06.021",
+	url = "http://www.sciencedirect.com/science/article/pii/S0045782516306077",
+	author = "Matthias Kabel and Felix Ospald and Matti Schneider",
+	keywords = "Composite materials, Homogenization, Hyperelasticity, Laminates, Lippmann–Schwinger equation, FFT",
 }
 @ARTICLE{HerzogOspald2016,
-  author = {Herzog, Roland and Ospald, Felix},
-  title = {Optimal Experimental Design for Linear Elastic Model Parameter Identification of Injection Molded Short Fiber-Reinforced Plastics},
-  journal = {International Journal for Numerical Methods in Engineering},
-  issn = {},
-  doi = {},
-  keywords = {optimal experimental design, anisotropic linear elasticity, short fiber-reinforced plastics},
-  year = {2016},
-  note = {NME-Apr-16-0258}
+  Title                    = {Parameter Identification for Short Fiber-Reinforced Plastics using Optimal Experimental Design},
+  Author                   = {Herzog, Roland and Ospald, Felix},
+  Journal                  = {International Journal for Numerical Methods in Engineering},
+  Publisher                = {John Wiley \& Sons, Ltd}, 
+  Year                     = {2017},
+  Doi                      = {10.1002/nme.5371},
+  ISSN                     = {1097-0207},
+  Keywords                 = {optimal experimental design, anisotropic linear elasticity, short fiber-reinforced plastics},
+        volume                   = {110},
+        number                   = {8},
+        pages                    = {703--725}
 }
 @article{MoulinecSuquet1994,
   author = {Moulinec, H. and Suquet, P.},
@@ -134,7 +127,7 @@
[…]
