### 2019-02-04 fa678dc
updated references
