+++ joss/paper.bib
@@ -12,7 +12,7 @@
 	publisher={Trelgol Publishing USA}
 }
 @ARTICLE{Montgomery-SmithJackSmith2011,
-  title = {The Fast Exact Closure for Jeffery’s equation with diffusion},
+  title = {{The Fast Exact Closure for Jeffery’s equation with diffusion}},
   journal = {Journal of Non-Newtonian Fluid Mechanics},
   volume = {166},
   number = {7-8},
@@ -36,7 +36,7 @@
   volume={54},
   number={6},
   doi={10.1007/s00466-014-1071-8},
-  title={Efficient fixed point and Newton-Krylov solvers for FFT-based homogenization of elasticity at large deformations},
+  title={{Efficient fixed point and Newton-Krylov solvers for FFT-based homogenization of elasticity at large deformations}},
   publisher={Springer Berlin Heidelberg},
   keywords={Composite materials; Finite deformations; Lippmann-Schwinger equation; FFT; 74B20; 45G10; 65T50},
   author={Kabel, Matthias and Böhlke, Thomas and Schneider, Matti},
@@ -45,7 +45,7 @@
 	file	= {555:Kabel_Böhlke_Schneider__Efficient_fixed_point_and_Newton-Krylov_solvers_for_FFT-based_homogenization_of_elasticity_at_large_deformations.pdf:PDF}
 }
 @article{KabelMerkertSchneider2015,
-	title = "Use of composite voxels in FFT-based homogenization",
+	title = {{Use of composite voxels in FFT-based homogenization}},
 	journal = "Computer Methods in Applied Mechanics and Engineering",
 	volume = "294",
 	pages = "168 - 188",
@@ -58,7 +58,7 @@
 }
 @ARTICLE{Kabel2016,
   author="Kabel, Matthias and Fliegener, Sascha and Schneider, Matti",
-  title="Mixed boundary conditions for FFT-based homogenization at finite strains",
+  title={{Mixed boundary conditions for FFT-based homogenization at finite strains}},
   journal="Computational Mechanics",
   year="2016",
   month="Feb",
@@ -130,8 +130,8 @@
   url={}
 }
 @article{Willot2015,
-  title = "Fourier-based schemes for computing the mechanical response of composites with accurate local fields ",
-  journal = "Comptes Rendus MÃ©canique ",
+  title = "Fourier-based schemes for computing the mechanical response of composites with accurate local fields",
+  journal = "Comptes Rendus Mécanique",
   volume = "343",
   number = "3",
   pages = "232 - 245",
@@ -140,7 +140,7 @@
   issn = "1631-0721",
   doi = "10.1016/j.crme.2014.12.005",
   url = "http://www.sciencedirect.com/science/article/pii/S1631072114002149",
-  author = "FranÃ§ois Willot",
+  author = "François Willot",
   keywords = "\{FFT\} methods",
   keywords = "Homogenization",
   keywords = "Heterogeneous media",
