### 2019-02-04 #1027 @fospald
@jedbrown I updated the citations (DOIs and titles) but there were no inline citations.
