### 2019-02-04 #1027 @jedbrown
`original method by [@MoulinecSuquet1994]` should be `original method by @MoulinecSuquet1994` (it will place the parentheses differently). There are several other citations like this.

If it's a part of the sentence, then it's "inline". The alternative would be to write so the citations are parenthetical, as in `original method [@MoulinecSuquet1994] and more recent ... on rotated [@Willot2015] and staggered [@SchneiderOspaldKabel2015] grids."
