### 2019-02-04 #1027 @fospald
@jedbrown Understood. I updated the citations.
