+++ README.md
@@ -2,7 +2,12 @@
   <a href="LICENSE" alt="GPLv3 license"><img src="https://img.shields.io/badge/license-GPLv3-brightgreen.svg" /></a>
   <a href="#" alt="no warranty"><img src="https://img.shields.io/badge/warranty-no-red.svg" /></a>
   <a href="http://joss.theoj.org/papers/add4f02dcedfaae8f81da200000d64ee"><img src="http://joss.theoj.org/papers/add4f02dcedfaae8f81da200000d64ee/status.svg"></a>
+  <a href="https://zenodo.org/badge/latestdoi/140197063"><img src="https://zenodo.org/badge/140197063.svg" alt="DOI"></a>
 </p>
+
+
+<a href="https://zenodo.org/badge/latestdoi/140197063"><img src="https://zenodo.org/badge/140197063.svg" alt="DOI"></a>
+
 
 # fibergen
 
