### 2019-02-04 #1027 @fospald
@jedbrown The DOI is: 10.5281/zenodo.2556861
for the v1.0 release.
https://doi.org/10.5281/zenodo.2556861
Not sure, why it is not found (takes time to register?)
