### 2019-02-05 #1027 @jedbrown
No need to let that hold up acceptance. You can upgrade to an annotated tag on GitHub at your leisure, if you so desire. (It makes `git describe` work and offers better provenance.)
