### 2019-02-05 #1027 @jedbrown
@openjournals/joss-eics Over to you. Thanks, @tisaac and @ctdegroot for reviewing.
