### 2019-02-05 #1027 @jedbrown
If you want an annotated tag on the server, you'd have to delete it and re-push, something like
```
git push origin :refs/tags/v1.0
git push --tags origin
```
