### 2019-02-06 #1027 @kyleniemeyer
It looks like that missing DOI is a fluke, so we can ignore that and proceed.
