### 2019-02-06 #1027 @kyleniemeyer
@fospald congrats, your article is now published in JOSS! Thanks for submitting a solid article and piece of software.

@tisaac & @ctdegroot thank you for reviewing this submission, and thanks to @jedbrown for editing it.
