### 2019-02-06 #1027 @fospald
@kyleniemeyer @jedbrown Looking at some published papers (including mine)
http://joss.theoj.org/papers/10.21105/joss.01027
http://joss.theoj.org/papers/10.21105/joss.01168
it seems in the section "Cite as:" (on the left) the first name is missing.
"Ospald, (2019)."
Can this be fixed?
