### 2019-02-06 #1027 @jedbrown
@fospald I believe that is the intended formatting, but it's unusual that the accepted papers list shows a URL `https://fospald.github.io/fibergen/` instead of the usual `fospald / fibergen`. http://joss.theoj.org/papers/accepted
